+++ src/Wrkzg.Frontend/src/App.tsx
@@ -10,6 +10,7 @@ import { Polls } from "./pages/Polls";
 import { Raffles } from "./pages/Raffles";
 import { Timers } from "./pages/Timers";
 import { Counters } from "./pages/Counters";
+import { Quotes } from "./pages/Quotes";
 import { SpamFilter } from "./pages/SpamFilter";
 
 export default function App() {
@@ -47,6 +48,7 @@ export default function App() {
         <Route path="raffles" element={<Raffles />} />
         <Route path="timers" element={<Timers />} />
         <Route path="counters" element={<Counters />} />
+        <Route path="quotes" element={<Quotes />} />
         <Route path="spam-filter" element={<SpamFilter />} />
         <Route path="settings" element={<Settings />} />
       </Route>